+++ webapp-frontend/src/app/anr-xapp/anr-xapp.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-anr-xapp',
+  templateUrl: './anr-xapp.component.html',
+  styleUrls: ['./anr-xapp.component.scss']
+})
+export class AnrXappComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}