+++ webapp-frontend/src/app/services/ac-xapp/ac-xapp.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { AcXappService } from './ac-xapp.service';
+
+describe('AcXappService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: AcXappService = TestBed.get(AcXappService);
+    expect(service).toBeTruthy();
+  });
+});