+++ docs/release-notes.rst
@@ -28,6 +28,8 @@ Version 1.0.3, 20 May 2019
 * Update E2 manager client to spec version 20190515
 * Add get-version methods to all controllers
 * Add simple page footer with copyright and version
+* Add AC and ANR xApp services
+* Rename signal service to E2 Manager service
 
 Version 1.0.2, 13 May 2019
 --------------------------