+++ docs/release-notes.rst
@@ -20,12 +20,13 @@
 RIC Dashboard Release Notes
 ===========================
 
-Version 1.0.3, 20 May 2019
+Version 1.0.3, 21 May 2019
 --------------------------
 * Add AC xapp controller
 * Add RAN type radio selector to connection setup
 * Update ANR xApp client to spec version 0.0.5
 * Update E2 manager client to spec version 20190515
+* Update xApp manager client to spec version 0.1.3
 * Add get-version methods to all controllers
 * Add simple page footer with copyright and version
 * Add AC and ANR xApp services